refactored UI-specific template code to TemplateFactoryUI
